# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Charter Communications NodeJS API project to demonstrate Swagger Documentation",
        "version": "1.0.0",
        "title": "Charter Communications API",
        "contact": {
            "email": "rajindarreddy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:3081",
    "basePath": "/api",
    "paths" : {
   

        "/networks" : {
            "get" : {
                "summary" : "Get all the networks",
                "description": "Get all the networks",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/networksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/network" : {
            "post" : {
                "summary" : "Save the network",
                "description": "Save the network",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "network object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "network" : {
                                    "type": "object",
                                    "$ref": "#/definitions/Network"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/networksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/network/" : {
            "put" : {
                "summary" : "Update the networks",
                "description": "Update the networks",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "network id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "network object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "network" : {
                                    "type": "object",
                                    "$ref": "#/definitions/Network"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/networksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/network/{id}" : {
            "delete" : {
                "summary" : "Delete the network",
                "description": "Delete the network",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "network id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/networksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },

        "/packages" : {
            "get" : {
                "summary" : "Get all the package",
                "description": "Get all the packages",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/networksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/package" : {
            "post" : {
                "summary" : "Save the package",
                "description": "Save the package",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "package object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "package" : {
                                    "type": "object",
                                    "$ref": "#/definitions/Package"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/packagesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/package/" : {
            "put" : {
                "summary" : "Update the packages",
                "description": "Update the packages",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "package id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "package object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "package" : {
                                    "type": "object",
                                    "$ref": "#/definitions/Package"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/packagesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/package/{id}" : {
            "delete" : {
                "summary" : "Delete the package",
                "description": "Delete the package",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "package id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/packagesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },


        "/shows" : {
            "get" : {
                "summary" : "Get all the show",
                "description": "Get all the shows",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/showsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/show" : {
            "post" : {
                "summary" : "Save the show",
                "description": "Save the show",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "show object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "show" : {
                                    "type": "object",
                                    "$ref": "#/definitions/Show"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/showsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/show/" : {
            "put" : {
                "summary" : "Update the shows",
                "description": "Update the shows",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "show id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "show object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "show" : {
                                    "type": "object",
                                    "$ref": "#/definitions/Show"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/showsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/show/{id}" : {
            "delete" : {
                "summary" : "Delete the show",
                "description": "Delete the show",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "show id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/showsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    }, 
    "definitions": {
        "networksResponse": {
            "type": "object",
            "properties": {
                "_id": {
                     "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
        "packagesResponse": {
            "type": "object",
            "properties": {
                "_id": {
                     "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "networks": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
               },
                "createdBy": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
        "showsResponse": {
            "type": "object",
            "properties": {
                "_id": {
                     "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "imdbRating": {
                    "type": "integer"
               },
                "createdBy": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
        "Network": {
            "type": "object",
            "properties": {
                "network": {
                    "type": {
                        "name": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "createdBy": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "Package": {
            "type": "object",
            "properties": {
                "package": {
                    "type": {
                        "name": {
                            "type": "string"
                        },
                        "networks": {
                            "type": "array"
                        },
                        "price": {
                            "type": "integer"
                        },
                        "createdBy": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "Show": {
            "type": "object",
            "properties": {
                "show": {
                    "type": {
                        "title": {
                            "type": "string"
                        },
                        "network": {
                            "type": "string"
                        },
                        "imdbRating": {
                            "type": "integer"
                        },
                        "createdBy": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }

        }
    }
}